own util header for request helper functions
