--- responseFiles/Upload/ft_putchar.c
@@ -1,14 +1,0 @@
-#include <unistd.h>
-
-void	ft_putchar(char c)
-{
-	write(1, &c, 1);
-}
-
-// int	main(void)
-
-// {
-// 	ft_putchar('y');
-// 	return (0);
-// }
-------WebKitFormBoundaryn9ieiOYTfyWbye6l--
